requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
discord.py==2.1.0
eventemitter==0.2.0
firebase-admin==6.0.1
frozenlist==1.3.3
gcloud==0.17.0
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
jws==0.1.3
msgpack==1.0.4
multidict==6.0.2
oauth2client==3.0.0
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
pyee==8.2.2
PyJWT==2.6.0
pyparsing==3.0.9
pyppeteer==1.0.2
python-jwt==2.0.1
requests==2.28.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.12
websockets==10.4
yarl==1.8.1
zipp==3.10.0
